Add check for fetch response status when fetching assets (#6522)
**User-Facing Changes**
Not worth being mentioned in release notes

**Description**
~We were rejecting any asset uri that was not a valid URL. However, the
Desktop app is able to resolve assets given as file path (e.g.
`/home/foo/mesh.dae`). This patch changes the behavior to only throw an
invalid URL error if it's not a desktop app.~

~Fixes #6519~

Update: We agreed on leaving it as it was and not support invalid URIs.
This PR now only adds an additional check for the fetch response status.
Former-commit-id: 9e1037a025a6ec01575814793cf4826d499a4ac9