Move sharedPanelState to CurrentLayoutContext (#6492)
**User-Facing Changes**
None

**Description**
This change moves the `sharedPanelState` from the `PanelStateContext`
into the `CurrentLayoutContext` alongside `selectedLayout`. This reduces
an awkward coupling that had to happen between the two contexts by
moving the shared state tracking for panels alongside the logic that
tracks current layout updates. Architecturally the
`CurrentLayoutContext` is a place for tracking the layout workspace and
panel related config and state. By moving this state into the
`CurrentLayoutContext`, the logic for determining when to clear the
`sharedPanelState` runs when specific actions on the layout require this
state to change.

Former-commit-id: 722326cb0ea64cf97a916ca6f5b4245078ffcfbd