Hoist memoization handling to react components (#6554)
This change updates the `MessagePathInput` and
`useCachedGetMessagePathDataItems` logic to use explicit memoization
rather than implicit memoization from some of the internal functions it
invokes. Explicit memoization is the more common pattern within our
react components and avoids bespoke memoization practices like that of
`messagePathStructures` which manually tracked `lastDatatypes` but could
be used in settings where datatypes are filtered down to a different
set. I find this explicit memoization easier to reason about and
optimize.

In https://github.com/foxglove/studio/pull/6544 I noticed that there was
only one non-explicit memoization user of `enumValuesByDatatypeAndField`
left and decided to change the call sites for `getMessagePathDataItems`
to accept the already computed enum values. This also allowed me to
remove the implicit memoization of `messagePathStructures` which was
manually managing its memoization via `lastDatatypes` and
`lastStructures` which would override if the inputs change like they
might when computing different sets of `relevantTopics` and
`relevantDatatypes` in `useCachedGetMessagePathDataItems`.

I've deleted the `topicsByName` selector and moved that to the two
places where it is used via `useMemo`. Since all that remains in
`selectors.ts` are two functions to deal with enums I've renamed it to
`enums.ts`.

I've added an `unwrap` function in `den/monads` to use in situations
where you want to assert that a value is defined and use that value
(usually in tests). This produces nicer stack traces than when you
attempt to use some undefined value later.

Former-commit-id: 9930d52d44ec987ca3ab9a94f067dfd29f2858b9