Detect console errors during integration tests (#6556)
Add an exit delay to integration tests for web and desktop allowing for
more rendering to happen and detect possible console errors that will
fail the integration test.

Without this timeout, an integration test will succeed as soon as it
sees the "App rendered" log which happens early in the rendering of the
app. As the app continues rendering there might be other errors that
happen so this change adds a 2 second timeout to allow those console
errors to fail the integration test.

Former-commit-id: 4e108efd15b53a777db655783a03c15ae80fd4ca