Reset last received messages when player ID changed (#6569)
**User-Facing Changes**
Fix messages from previous connection influencing visualizations

**Description**
Resets the last received messages when the player ID changes. 
For the foxglove-websocket player, the player instance remains the same
during re-connections, but the ID changes. In that case, we also want to
reset the last received messages which stem from a previous connection.

---------

Co-authored-by: Roman Shtylman <roman@foxglove.dev>
Former-commit-id: 2adc2704ddb5d1db4e8b443290ef35ec56c8c4d7