Improve interactivity of plot series path editing (#6577)
Improve interactivity of plot series path editing.

Currently when a user is actively editing a series path in the plot
panel the paths that are inputs to `usePlotPanelData` are rapidly
changing and this causes a complete rebuild of the datasets on each
keystroke which can be very slow on larger datasets. This patch tries to
eliminate most of that work in the cases where it's not useful by
filtering out changes to series paths that will not result in any new
messages being processed and included in the dataset.

This involves pre-filtering paths that either are not valid ros paths or
that parse to a topic that is not present in our dataset. And the logic
that builds datasets from these filtered paths tries to eliminate any
changes that wouldn't result in any new messages.


Former-commit-id: cbe2b057e7eafa7ec45459e8a429e8996470f863