Feature: Add optional RGB rendering of 16 bit single channel images (#6570)
From previous PR here: https://github.com/foxglove/studio/pull/6123
**User-Facing Changes**

Adds a setting to the image panel to allow mono 16 images to be rendered
across a color space instead of as a grayscale.

**Description**

A boolean RGB setting to the image. When turned on, instead of rendering
as a grayscale black -> white, it renders as an HSV blue -> red. This
allows easier differentiation between similar but different values.

This setting needs to be added to the `docs` page here:
https://foxglove.dev/docs/studio/panels/image#settings.


![image](https://github.com/foxglove/studio/assets/23124304/48e4ee2b-0841-49fd-b170-db54ee77d23c)gi

---------

Co-authored-by: Ross Lannen <ross.lannen@gmail.com>
Former-commit-id: 40cddd83aac2f89faaac37db245a94c6135abe06