Fix topic stats disappearing when filtering the topic list (#6718)
**User-Facing Changes**
Not worth calling out in release notes

**Description**

Fixes FG-4839

The DirectTopicStatsUpdater didn't handle when rows were dynamically
added or removed from the topic list. This PR fixes the issue by using a
MutationObserver to see when nodes are added/removed from the parent
element to trigger updates.

This is also a must-have for #6717 because we're going to virtualize the
list using `react-window`.

Former-commit-id: fe4f681330bdf31be6dc00ae914d25add54b38fb