Fix extra lines rendering for LinePrimitives (#6805)
**User-Facing Changes**
<!-- will be used as a changelog entry -->
- Fix occasional extra lines in SceneUpdate line primitives

**Description**

Fixed instanceCount for `LINE_LOOP` primitive types. Because we add one
to the `numVertices` when we have a line loop, the number of lines in a
line loop should be `numVertices - 1`, or the original number of points
on the primitive. The story I added has one square defined with 4 points
(not closed) and the other defined with 5 points (already closed). We
then add one in calculating `numVertices` because we close the loop in
`serializePositions` even if the loop is already closed. This makes the
number of lines (`instanceCount`) equal to 4 for the first and 5 for the
second, whereas previously it was 5 and 6. This would cause the
renderable to read additional points from existing, larger
`positionBuffers` that were used after an update.

<!-- link relevant GitHub issues -->
Fixes: FG-4944
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->
Former-commit-id: f17c30d467e92429c807ebf1ec93d0a75206e6d9