Renderer uses `SceneExtensionConfig` (#6848)
**User-Facing Changes**
<!-- will be used as a changelog entry -->
- none

**Description**

In order to support adding or overriding scene extensions in the panel I
have added a `sceneExtensionsConfigOverride` prop to the `ThreeDeePanel`
component, which is merged with a default config and passed to the
`Renderer`. The config contains `init`ializers that are callbacks which
return the new `SceneExtension` instance, these are treated generically
with exceptions for scene extensions that are assigned to render
variables on the `Renderer` itself.

Extensions are placed in the config by their extensionId. The
extensionId has been made a static member on the class so that it can be
more easily subclassed and overriden by `sceneExtensionOverrides`.

`ImageMode` input props were reduced to make more consistent with other
scene extensions. This also required `enableImageOnlyMode` (and disable)
to be made a public so that `ImageMode` can utilize it directly. I know
this is an additional public method, but because the initializers are
outside of `Renderer` they no longer have access to these methods to
begin with.



<!-- link relevant GitHub issues -->
FG-5027
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->

Former-commit-id: 9553ebdc23e9c18783fc2290700869134eee2325