Fix issue with disappearing preloaded plots (#6884)
**User-Facing Changes**
None.

**Description**
Plots that used block data were cleared when switching tabs. This PR
fixes that.

The problem was that `addBlock` could come before `register` had
completed, and thus `evictCache` would clear out the new block that just
got added. (Theoretically this is a timing problem but it's more
complicated than it seems at first glance; waiting for `register` to
complete before doing `addBlock` would not make sense, since the
`addBlock` mechanism in the rendering thread has already done the work
to determine that a client actually needs the block data.)

<!-- link relevant GitHub issues -->
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->

Former-commit-id: 451911d75371071c00aae2ef5efebaa2a989b229