Reduce console errors for LINE_LIST annotations with an odd number of points (#6887)
**User-Facing Changes**
None

**Description**
Previously a LINE_LIST annotation with an odd number of points would
spew console errors when it encountered NaNs inside
`computeBoundingSphere()`. Instead we now just ignore the extra point;
this matches how we handle ROS Marker line lists:


https://github.com/foxglove/studio/blob/4e55d46b07729465b9126ceba5e6ecf17affad9c [formerly f15de61e9cd4d92400731108c6f9bf5f3bf3b822]/packages/studio-base/src/panels/ThreeDeeRender/renderables/markers/RenderableLineList.ts#L81-L83

I've added a story to show that we can render these, but note that the
story does not capture the console.error behavior so it wouldn't have
detected this issue on main.
Former-commit-id: 8c6af5957983ef7a90945912f7d245ee3eed6570