Repair string references in the plot data worker (#6915)
**User-Facing Changes**
None.

**Description**
We were using a fair amount of memory storing multiple copies of strings
in the plot worker, because `structuredClone()` creates new strings _for
each reference_ to a string. This PR adds a new library (`strPack`) that
combines references to identical strings.

It's not a perfect solution. I also did not touch anything in the main
thread despite similar issues. But this attempt alone shaves off 20% of
our memory usage in the worker, which is a pretty serious gain for what
ultimately was not that much work.

<!-- link relevant GitHub issues -->
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->

Former-commit-id: 66ba8ae49d5bd57257eee23a9720f40894c216d0