Avoid adding the same ID twice to subscriberIdsByTopic (#7027)
**User-Facing Changes**
Fix duplicate messages being sent to a panel if it subscribed twice to
the same topic.

**Description**
This change updates the `MessagePipeline` store `updateSubscriberAction`
function to avoid adding a subscriber ID multiple times to the
`subscriberIdsByTopic` array for a particular topic. Including the ID
multiple times results in other logic in the store which handles message
distribution to erroneously provide the message to the panel multiple
times here:
https://github.com/foxglove/studio/blob/main/packages/studio-base/src/components/MessagePipeline/store.ts#L289-L300

A panel might subscribe to the same topic multiple times if it an array
of subscription requests and includes the same topic multiple times and
does not itself de-duplicate the array entries based on topic name. A
panel is not required to do such de-duplication.

This issue was uncovered by #6984 which created a situation where the 3d
panel thought it could subscribe to the same topic twice because the
original topic was a `foxglove.Grid` schema and there was a converter
installed which provided a conversion from `foxglove.Grid` to
`foxglove.PosesInFrame`. Both of these schemas are 3d renderable and the
panel wanted to subscribe for both.

Fixes: #6984
Former-commit-id: 1d741d2e471e1f4e244c2e92c86fe77823dbefea