Convert UserScriptStateContext to zustand store (#7048)
When scrubbing or seeking, the user scripts are reset which causes the
`UserScriptStateContext` _state_ to get updated. Because the entire
context is not selectable this invalidates the _value_ of the context
which causes it to render and any downstream users of the context to
re-render. This re-renders a lot of things downstream that don't need to
render because they only need the stable actions functions and not the
state.

This change updates the context to use zustand and allow for separately
selecting the actions and the state.

Former-commit-id: f6fa31a6e228c8d74773aa7498a254e6a145329f