Use a more specific workspace selector in AppBar (#7055)
When selecting from the workspace context using a selector, it is
important
to avoid having the selector value change if the value you want to
select is not changing. This helps reduce
renders of unrelated components when other aspects of the selectable
context change.

This change updates the AppBar and AppMenu selection of the left/right
sidebar open state to be more stable
thus avoiding re-rendering those components when other aspects of the
workspace context change like sidebar sizes.

Former-commit-id: 41f583ead9dafa9406a1105b483530ea9922e177