Fix a range of issues with user scripts and plots (#7041)
**User-Facing Changes**
Fixed a handful of issues where plotting user script topics could lead
to broken plots.

**Description**
This one is a doozy. It fixes three separate bugs that combined to
create the poor experience described in [this GitHub ~issue~
discussion](https://github.com/orgs/foxglove/discussions/855):
1. If block data already exists on a topic a plot panel needs when it's
mounted, the main thread delivers block data to the plot worker before
the client is registered. The plot worker would drop the data.
a. **Solution:** Now we hold on to incoming block data in anticipation
of the registration of a client.
3. The plot panel block processor treated empty arrays on a topic as
valid message data, not as being empty. This can occur if the
`UserScriptPlayer` is partway through processing data when a
subscription to a user script topic is cancelled.
  a. **Solution:** Empty message arrays now indicate an empty block.
5. The `BlockLoader` does not handle changes in subscriptions
gracefully, particularly for user script topics.
a. **Solution:** If topics change while we're loading data, we return
from `load()` and wait until the next iteration to continue.

<!-- link relevant GitHub issues -->
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->

Former-commit-id: c485bc8026fb8c1f0ba48074b16fb7bd255d21a6