Update comment for BlockLoader early bail (#7066)
**User-Facing Changes**
None

**Description**
Updated the comment in BlockLoader to be more clear on what the logic is
trying to do. Reference:
https://github.com/foxglove/studio/pull/7041/files#r1380461856

I've updated the equality to `isEqual` from lodash because we were
already using that in this file.

---
What I don't understand reading the original issue is why this logic
needed to be added. Block loader #topics are only set via `setTopics`
which also invokes an abort controller. This abort controller _should_
be triggering this bail logic a few lines above the newly introduced
logic:

```
// No results means cursor aborted or eof
if (!results) {
  await cursor.end();
  return;
}
```

Seems like that was not happening? I'd like to understand what
specifically was violated.
Former-commit-id: 0e3b41fe22e71a423c39750959d7203be9855cd9