Fix some Tab panel bugs (#7075)
**User-Facing Changes**
Fixed issues where splitting tab panels would sometimes result in panels
losing their config, or multiple panels sharing the same config.

**Description**
- The logic that "clones" panels residing in tabs when the tab is split
lives in `getSaveConfigsPayloadForAddedPanel`. This function was only
being invoked when the panel being split was not inside a tab layout. It
needs to be invoked regardless of whether the panel being split is in
nested in another tab.
- When cloning panels while splitting, the `relatedConfigs` was being
used as a source for the newly cloned panels' configs. This
`relatedConfigs` was a subset of the overall `configById`, which led to
newly cloned panels sometimes having no config (because it could not be
found in `relatedConfigs`). I changed this to pass in the whole
`configById`.
- Removed the `relatedConfigs` concept entirely since we do not use this
in the app (it was related to a defunct "presets" feature).
- Removed some test helpers which were no longer used.

Former-commit-id: a517eaf3c93891c1e4531d62d5110be43749e2fd