Fix deserialization errors hiding errors from other topics (#7077)
**User-Facing Changes**
None

**Description**

The IterablePlayer uses the `connectionId` as part of the key in the
problem manager. The source was not sending any `connectionId`, so
errors on multiple channels would override each other and only one was
visible at a time. With this change, each channel's latest error is
visible.


https://github.com/foxglove/studio/blob/081721726153e6cc7b3d070b5dc8bd9d7ee3a875 [formerly 3f87e3009acfb94020ad8270492225179ca6a5d3]/packages/studio-base/src/players/IterablePlayer/IterablePlayer.ts#L652

Fixes FG-5631
Former-commit-id: 5df23212f1bb96c430bbe43d567b4331e0d0b0b4