Stabilize the bounds reference for state transition panel (#7142)
When playing recorded data, the state transition panel would invalidate
the dataset bounds reference for every frame. This causes unecessary
work downstream.

This change updates the databounds memo to use a stable fixed bounds
when the inputs for those bounds are not changing but the
`currentTimeSinceStart` is changing but not used for the current
configuration.

Former-commit-id: f57d80ce7f34a791612c3a3dfbdcd8ecb33e0a0a