Update subscriptions on topic alias changes (#7153)
This change fixes a behavior where changes to global variables would not
cause updated subscriptions for the wrapped player. This would happen
because aliased subscriptions were only updated in the
`setSubscriptions` call. However, for some panel workflows they would
not re-subscribe when the topics list changed because they were already
subscribed to the topic they wanted (even though it was not yet present
in the topic list). Plot panel is an example of such a panel. Because
they had already specified their subscriptions there was no additional
call to `setSubscriptions` and no re-calculating of aliased
subscriptions for the original topic.

This change updates the `TopicAliasingPlayer` logic to re-process
aliased subscriptions and call `setSubscriptions` on the wrapped player
for all three workflows: setGlobalVariables, setSubscriptions, and
player state updates. Now if a panel had previously subscribed that
subscription will be properly aliased.

I've also changed the `setSubscriptions` behavior. Previously it would
not set subscriptions until topics were received from the underlying
player. I see no reason to keep this restriction since alias functions
could be hard-coded and not require any input topics.

Former-commit-id: af441adaf38a67380a73b21d680fc1ce9051af62