Hide individual channel override in log settings for warn and error (#7156)
When setting the studio log level to warn or error, the individual log
file overrides (to hide a speciifc file's logs) does not work. This is
by design so warning and error logs are always shown.

This change updates the log settings sidebar to hide the individual file
names when "warn" or "error" log level is selected.

Former-commit-id: 38bdb07b1fbe601dec17e0e588da9ebb41fad46d