Flaky plots IV: A New Hope (#7183)
**User-Facing Changes**
N/A

**Description**
I'm continuing my quest to fix our flaky stories. In this attempt I:
1. Changed all Plot and State Transition stories to just use single
light mode plots (where applicable), since it seems that some of our
problems stem from timing discrepancies between the two. In general it's
nice to see both themes but in practice this seems to still cause us
problems. This could just be temporary.
2. Fixed an ordering issue in the `Chart` component. It turns out that
the previous `queueUpdate` logic was incorrect for specifically the
first `update` call: if updates were queued while we were still
initializing, we would incorrectly push the first `update` on top of
them. Instead, we insert the first update at the _beginning_ of the
queued updates.

Former-commit-id: 1e390c58831cb45d56f553487a053dee5565cbf2