Fix Image panel images not clearing when looping playback (#7173)
## User-facing changes

 - Fix Image panel images not clearing when looping playback


## Description

Image renderables weren't being cleared on seek for two reasons. The
first being that we weren't queuing an animationframe once the timeout
is up, and the second being that if the next image message would fail
decoding, then we would clear the timeout so the renderable would still
have a stale final image being displayed. To fix these, I've added a
queueAnimation frame in the timeout and now when erroring we show an
error image instead of the previous frame.

Fixes: FG-5772
Former-commit-id: da8de7d04f33ce346b1f62d410a8299b9ffdcb37