[Image Panel] Fix unmatched frameId error when switching to none from non-existing topic (#7189)
## User-facing changes

- [Image Panel] Fix unmatched frameId error when switching to `None`
from non-existing topic

## Description

When loading a config where the image topic is defined but the
calibration topic is not, the image is read and decoded, but a fallback
camera model is not created onDecode because the calibration topic is
not `None`. Then when the user switches to `None`, a new fallback is not
created because it does not

This issue existed because we were only creating fallback calibrations
on image decode. If an image had been received when there was a
non-existing calibration topic, a fallback is not created. When
switching to `None` as a calibration topic we were getting an error
because the fallback calibration had not been created. This fixes the
issue by creating a fallback calibration when switching to `None` if an
image renderable already exists.




Fixes: FG-5796
Former-commit-id: 97f91080d26a5f2ab3ecfa82e77421a71dd06193