Update point clouds to default to `z` instead of first field (#7206)
**User-Facing Changes**
<!-- will be used as a changelog entry -->
- Point clouds will default to color by `z` when intensity and rgb
fields aren't present

**Description**

Updated to color by `z` as a fallback for all colorMode applicable
renderables. Also in this I consolidated the autoselecting of color
settings, and updated all renderables to use the same method. I also
consolidated the setting of fields for pointlcouds since `isSupported`
would run on both foxglove and ros pointclouds, and we should be using
it as a flag to determine whether to update the autoselection of color
settings aside from the colorField being `undefined` because the
colorField can be undefined for `rgba-fields`.

We're also defaulting to `x` in `colorMode.ts`
[here](https://github.com/foxglove/studio/blob/main/packages/studio-base/src/panels/ThreeDeeRender/renderables/colorMode.ts#L199),
which we use for FoxgloveGrid, and Images. ~~Should I update that to use
`z` as well?~~ updated.


<!-- link relevant GitHub issues -->
Fixes: FG-5831
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->
Former-commit-id: 64d331b20f18f29cc6f16b1f936397feae11d372