Better downsampling for the State Transition panel (#7212)
**User-Facing Changes**
Introduce a new downsampling algorithm for the State Transition panel to
greatly improve its performance.

**Description**
This PR introduces a downsampling algorithm specifically for state
transition data called `downsampleStates`. It is conceptually similar to
our existing `downsampleTimeseries` algorithm, which divides the
viewport into a set of equally-sized intervals.

The nature of state transitions lets us do something a bit more
sophisticated: when an interval contains more points than are in our
"budget" (put differently: it contains more than one distinct state
transition), we render it in gray. We can work on the UX of this, but it
does work, includes a suite of tests, and resolves the high-level issue
with the panel's rendering performance.

Here is an example from nuScenes. This scenario intentionally contains a
lot of state transitions, more than might actually occur in everyday
use; it's just a good way of demonstrating how this might work.


https://github.com/foxglove/studio/assets/4588553/a9b3e86b-b673-4529-bd8d-c126a8eb34c4

When the user mouses over a segment with multiple states, we render them
in the tooltip:
<img width="917" alt="image"
src="https://github.com/foxglove/studio/assets/4588553/bca84d9f-7b7a-4ddf-a99e-2fd7ac4cc36f">
Former-commit-id: fb267b80d61005b8e28395711774975c86476d1e