Use a stable empty array for IterablePlayer (#7268)
When the iterable player produces a state update, it "consumes" the
outgoing message array and resets the outgoing messages to an empty
array for the next state update. It would make a new empty array every
time which changes the referential equality of `activeData.messages` and
makes it look like there are new messages. Since state is emitted for
block loading as well as playback, downstream consumers would think the
current frame messages have changed.

This change updates the logic to use a stable EMPTY_ARRAY so it does not
change with every emit.

Former-commit-id: 66979ee336e2add1e59ad0aa38d65c4297dbb3de