Lower maximum frame size in websocket player (#7303)
**User-Facing Changes**
<!-- will be used as a changelog entry -->
- Reduce OOM crashing on live datasources when application is slow or
under an inactive tab


**Description**

Lower max size of single frame in websocket player from 1500MB to 600MB.
Initially my thinking was to make this about the same as the
Blockloading cache limit (2GB, at the time). However, that thinking
wasn't quite correct because there is only ever really one instance of
the block data, but in this case there could be a current frame being
processed that's 1.5GB and a frame being built up that hasn't been
emitted that's 1.5GB. This on top of the normal memory usage from panels
would be enough to kill apps. So I lowered it to something that seems a
bit more reasonable when taking into account that two frames could exist
at a time that are this size.

When studio is too slow to emit a frame from the websocket worker,
messages accumulate in the next frame. The larger the next frame grows,
the more time it could take to process the next frame, thus causing the
following frame to also grow larger, etc. This change doesn't fix this
issue but should reduce OOMs when there is an accumulation of data in
panels as well as in frames, and keep the UI responsive and processing
messages so as to not build up memory pressure more.


<!-- link relevant GitHub issues -->
FG-6238
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->

Former-commit-id: 3cb303aeaf4487e6a2f3ddf7a35599fbb334939f