Fix Image panel memory leak when disposing while decoding an image (#7310)
**User-Facing Changes**
<!-- will be used as a changelog entry -->
- Fix Image panel memory leak when unmounting while decoding an image

**Description**

Terminating the worker while the promise hasn't been resolved results in
the promise never getting resolved nor rejected. This promise is
referenced within the renderable itself and has callbacks within the
renderable. This promise being linked to the global worker object and
the renderable cause the entire renderer to not be disposed, leaving
multiple instances in memory. This abort lets the promise resolve and
clean itself up so that it doesn't keep it's associations in memory.

Easy way to test: 
Source with image (preferably large), and layout with image panel
reading image.
Start playback and either close the panel while playback is happening or
change the transform preloading settings (this reinstantiates the
renderer).
Close the panel. 
Take a heap snapshot and see that there are multiple `Renderer`
instances.

<!-- link relevant GitHub issues -->
FG-6148
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->

Former-commit-id: 3f43e72e8d30d2997b665051dccb19835675c47f