Keep referentially stable blocks in TopicAliasingPlayer (#7338)
The AliasingStateProcessor (part of the TopicAliasingPlayer machinery)
processes blocks to add the aliased MessageEvent arrays to
`messageByTopic`. The existing behavior would receive a new set of
blocks and update the aliased topic messages in all of them. This
resulted in a new messages array reference for all aliased topics any
time a new block is added. Changing the message array reference signals
to downstream that the specific messages in the block for that topic
have changed and can trigger work on those messages. In the case of
plots, it causes the internal block cursors to think the data is changed
and re-send all the block data for the series.

This change updates the `AliasingStateProcessor` with a new assistant -
the `BlockTopicProcessor`. The BlockTopicProcessor remembers the aliased
message arrays for the aliased topics and returns those if the source
topic messages are unchanged (referentially stable). This allows
downstream consumers to continue using the referential equality of
message arrays to identify if the data is changed.

Former-commit-id: f683ba2f0544e44e8bdaa7af8f0536ae5a265d6e