Periodically trigger render when processing blocks (#7345)
When a signal is added to the plot, the main thread of the panel
processes the blocks to populate the worker with the series data. During
initial loading of the data source, the blocks are populated async which
provides an opportunity for the panel logic to stop processing blocks,
send the data to the accumulation worker, and trigger a plot render.

However, when the data source is already loaded with the topic we want
to plot (like when you split a plot), no such natural async opportunity
arrises and the plot main thread attempts to process all the blocks into
one update call to the accumulation worker and then render - since all
the blocks are available. This large update call hangs the main thread
(building a massive postMessage) and the user gets no incremental
feedback that the plot is loading - it sits there empty and then seconds
later _bam_ it appears all at once.

This change introduces a new way for dataset builders to handle blocks.
The optional `handleBlocks` method of `IDatasetsBuilder` when
implemented by a builder is invoked with incoming blocks and a progress
callback. The builder can process its blocks and periodically call the
progress callback. In the coordinator, the progress callback queues up a
render and a wait. The wait is important as it creates an opportunity
for the main thread to dispatch the queued datasets and render. Thus
showing incremental progress on the plot. Without the delay, the blocks
would be processed as a tight loop with no incremental rendering.

Former-commit-id: 44074c51cd2ed49e38f75a1bc43d5ab7c94dcd28