Move parseRosPath to @foxglove/message-path (#7371)
Move `parseRosPath.ts` and `constants.ts` to a new module
`@foxglove/message-path`.

- Rename `parseRosPath` -> `parseMessagePath`
- Rename `RosPath` type -> `MessagePath`
- Rename `RosPrimitive` type -> `PrimitiveType`
- Rename `constants.ts` -> `types.ts` in the new module

Note: I've removed the _memoize_ wrapping for `parseMessagePath`. I
don't think this is the appropriate level to memoize things but I can
put it back if we want to minimize change. I don't know how we will
prove or disprove this is important to memoize.
Former-commit-id: 643ad257bea126ce23e4c130def786714439c7ac