Suppress webpack client overlay for webworker importScript errors (#7405)
When a webworker is terminated, any pending `importScript` calls are
cancelled by the browser. These appear in the devtools network tab as
"(cancelled)" and bubble up to the parent page as errors which trigger
`window.onerror`.

Webpack devserver sees these uncaught errors and shows the error
overlay. This is a false positive since the worker is already
terminated. This change suppresses the overlay in this scenario.

Former-commit-id: 3c40c9688bc299416071616a8daf43aa3c240aa7