Reduce downsamples and re-renders for plot (#7362)
Prior to this change, any plot panel would re-downsample its dataset on
any player state update regardless of whether it had new data to render
or not. This change updates the `handlePlayerState` method of
`IDatasetsBuilder` to return the number of datums added in addition to
the range. The coordinator can avoid downsampling if there are no datums
read.

This change also splits the downsample dispatch from the rendering of
the downsampled datasets. Since these are powered by separate workers we
can start work on a new downsample request while rendering the old.

Former-commit-id: 2841de2223eebb74af3d2f72fc7c2955ae88f32e