Define if it is production or development by argv on webpack (#170)
**User-Facing Changes**
After users have installed Lichtblick, the name of the application
became "Lichtblick Dev" instead of "Lichtblick".

**Description**
It happens because in the build process, it was expecting a environment
variable called RELEASE, and in it absence, it added " Dev" at end of
the name. The fix could be add this variable, but as the project has the
commands separated "desktop:build:dev" and "desktop:build:prod", it
would be more consistent to not depend on the environment but from the
argument passed in the command. The other webpack config files already
use the "mode" argument, instead of an environment variable.

**Checklist**

- [x] The web version was tested and it is running ok
- [x] The desktop version was tested and it is running ok
- [x] I've updated/created the storybook file(s)

Former-commit-id: c52e37bbfc1e7709e3b439d66f2c0998161c668f