[Fix] Suppress importScripts error as done in web version (#172)
**User-Facing Changes**
User receiving error about "importScripts" when running the desktop in
dev mode.

**Description**
This bug is related to certain calls being cancelled when a web worker
is terminated. While this error can be a false positive—since the
termination of a worker is usually not a concern—it can still involve
legitimate cancelled calls. Handling such situations should be managed
within the web worker's own logic. Regardless, the error will still
appear in the developer console if debugging is necessary.

The proposed fix is to suppress this error for the desktop version, just
as it was done for the web version. The user who reported the bug has
confirmed that the web version works without warnings.

**Checklist**
- [x] The web version was tested and it is running ok
- [x] The desktop version was tested and it is running ok
- [x] I've updated/created the storybook file(s)

Former-commit-id: fb85f0eb193b9ad61ba8192732068fc373b60b45