Removed integration-test folders from scanning (#217)
**User-Facing Changes**
N/A

**Description**
Removed integration tests from Sonar scanning.

Why not include E2E and integration tests in SonarQube:
Focus on production code: SonarQube is primarily designed to check the
quality of production code (not test code). It focuses on static
analysis, security, and code coverage metrics in unit tests, which are
more precise in testing the internal logic of the code.

Unit test coverage is more relevant: Code coverage is a key metric to
ensure that the production code has been well-tested in smaller units,
which is the focus of quality analysis. E2E and integration tests can
involve multiple layers of the application, which dilutes the
granularity of the coverage.

Unnecessary noise in metrics: Including E2E and integration tests can
artificially inflate the code coverage metric, as these tests tend to
cover large portions of the code but do not provide the same accuracy as
unit tests. This can lead to less clear reports.

Complexity in reports: E2E and integration tests may include many
irrelevant details for code quality analysis, which can complicate the
interpretation of SonarQube results.

---------

Signed-off-by: Luiz Bezerra <luiz.bezerra@ctw.bmwgroup.com>
Former-commit-id: 8c0808f5cff57a964f569fcba1e265232b2c3833