Refactor WorkspaceContent component and update i18n files
This commit refactors the WorkspaceContent component in the suite-base package. It removes the unused import of the `t` function from the `i18next` library and adds the `useTranslation` hook from the `react-i18next` library to handle translations. Additionally, it updates the i18n files for English, Japanese, and Chinese languages to include translations for the new "layouts" key.


Former-commit-id: e47cd51cc405c427f129779214214167c1e003c4