Fix toggling of panels by listen only to revert event (#264)
**User-Facing Changes**
Users use to experience some toggling in some layouts, which made
impossible to use Lichtblick.

**Description**
The toggling was solved by filtering the "change event" in the
CurrentLayoutProvider, and update in that specific moment only if it is
a revert action (revert the unsaved modifications on current layout)
instead of all changes. For some reason this change event is possibly
duplicated and is not necessary in this part.

The tests passed and it was tested manually too. It seems that all
functionality regards to the layouts is still working.

**Checklist**

- [x] The web version was tested and it is running ok
- [x] The desktop version was tested and it is running ok
- [x] Files constants.ts, types.ts and *.style.ts have been checked and
relevant code snippets have been relocated